Tidy up and rename KeyStageFilterable
